Fix some of various typing issues
which were randomly noticed and even more randomly
picked to get fixed
and let linters do their work in files affected

+++ src/api/models/MapVersion.py
@@ -13,11 +13,11 @@ class MapSize:
     width_px: int
 
     @property
-    def width_km(self) -> int:
+    def width_km(self) -> float:
         return self.width_px / 51.2
 
     @property
-    def height_km(self) -> int:
+    def height_km(self) -> float:
         return self.height_px / 51.2
 
     def __lt__(self, other: MapSize) -> bool:
